REP-202 refactor tests
