+++ cares-ui/test/reporter/models/searchResultsModel.test.js
@@ -52,65 +52,4 @@ describe('SearchResultsModel', () => {
       expect(formatedResult).toMatch(/1234567890/)
     })
   })
-
-  describe('validate', () => {
-    it('validates that the reporter passes all business rules', () => {
-      const model = new SearchResultsModel()
-      const options = {}
-      model.data = {
-        first_name: 'first',
-        last_name: 'last',
-        phone_number: 123
-      }
-      model.validate({}, options)
-      expect(options.error).toEqual(undefined)
-    })
-
-    it('does not validate the reporter when the question is empty', () => {
-      const model = new SearchResultsModel()
-      spyOn(model, 'getPanelByName').and.returnValue({
-        getElementByName: jasmine.createSpy('getElementByName').and.returnValue({
-          isEmpty: () => false
-        })
-      })
-      const options = {}
-      model.data = {
-        first_name: 'first',
-        phone_number: 123
-      }
-      model.validate({}, options)
-      expect(options.error).toEqual(undefined)
-    })
-  })
-
-  describe('loadJsonRules', () => {
-    it('loads the json rules from an api', (done) => {
-      const rule = {
-        "and": [1,2]
-      }
-      mockAxios.onGet(getBreRuleSetRoute('ReporterCreateScreenBusinessRules')).reply(200, {
-        rules: [{
-          name: 'rule',
-          logic: rule
-        }]
-      })
-      const model = new SearchResultsModel({})
-      model.loadJsonRules().then(() => {
-        const rule = model.engine.find((rule) => rule.identifier === 'rule')
-        expect(rule).toEqual(rule)
-        done()
-      })
-    })
-
-    it('sets the create reporter error when there is a problem with BRE', (done) => {
-      mockAxios.onGet(getBreRuleSetRoute('ReporterCreateScreenBusinessRules')).reply(404)
-      const createReporterError = jasmine.createSpy('createReporterError')
-      const props =  { createReporterError }
-      const model = new SearchResultsModel(props)
-      model.loadJsonRules().then(() => {
-        expect(createReporterError).toHaveBeenCalled()
-        done()
-      })
-    })
-  })
 })
